backend/pyproject.toml
[project]
name = "athena"
version = "0.1.0"
description = "RAG-powered research assistant with hybrid search and evaluation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "sentence-transformers>=3.0.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.3.0",
    "zhipuai>=2.1.0",
    "pypdf>=5.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-docx>=1.0.0",
    "python-multipart>=0.0.12",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.27.0",
]
eval = [
    "ragas>=0.2.0",
    "datasets>=3.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI Depends() in function defaults is intentional

[tool.ruff.lint.per-file-ignores]
"app/agents/*.py" = ["E501"]      # long prompt strings
"app/generation/prompts.py" = ["E501"]
"app/api/routes.py" = ["E501"]
"tests/*.py" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "pgvector.*",
    "ragas.*",
    "datasets.*",
    "sentence_transformers.*",
    "langchain_community.*",
    "langchain_anthropic.*",
    "langgraph.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["app.agents.*"]
disable_error_code = ["type-arg", "no-any-return", "assignment", "return-value"]

[[tool.mypy.overrides]]
module = [
    "app.models.orm",
    "app.ingestion.chunker",
    "app.ingestion.embedder",
    "app.ingestion.loader",
    "app.retrieval.*",
    "app.generation.prompts",
    "app.services.document_service",
    "app.api.deps",
    "app.api.routes",
    "eval.runner",
]
disable_error_code = ["type-arg", "no-any-return", "import-not-found", "unused-ignore", "arg-type", "no-untyped-call"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "eval*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
